+++ geotrellis-python/.gitignore
@@ -1,0 +1,4 @@
+build/
+dist/
+*.egg-info
+*.pyc
