+++ geotrellis-python/geotrellis/python/spray/json/package_scala.py
@@ -1,0 +1,6 @@
+
+class DeserializationException(Exception):
+    def __init__(self, msg, cause = None, field_names = []):
+        self.msg = msg
+        self.cause = cause
+
