+++ geotrellis-python/geotrellis/util/GetComponent.py
@@ -1,0 +1,10 @@
+class GetComponent(object):
+    def get(self):
+        pass
+
+def get_component(_get):
+    class TempGetComponent(GetComponent):
+        def get(self):
+            return _get
+    return TempGetComponent()
+
