+++ geotrellis-python/geotrellis/util/SetComponent.py
@@ -1,0 +1,10 @@
+class SetComponent(object):
+    def set(self):
+        pass
+
+def set_component(_set):
+    class TempSetComponent(SetComponent):
+        def set(self):
+            return _set
+    return TempSetComponent()
+
